Added the two subclasses for the lists
