+++ llitgi/ViewControllers/ViewControllerFactory.swift
@@ -28,28 +28,28 @@ final class ViewControllerFactory {
         return AuthorizationViewController(dataProvider: self.dataProvider, themeManager: self.themeManager)
     }
     
-    func instantiateLists() -> [BaseListViewController] {
-        let listViewController: BaseListViewController = self.instantiateList(for: .myList)
+    func instantiateLists() -> [ItemsViewController] {
+        let listViewController: ItemsViewController = self.instantiateList(for: .myList)
         listViewController.title = L10n.Titles.myList
         listViewController.tabBarItem = UITabBarItem(title: L10n.Titles.myList, image: #imageLiteral(resourceName: "list"), tag: 1)
         
-        let favoritesViewController: BaseListViewController = self.instantiateList(for: .favorites)
+        let favoritesViewController: ItemsViewController = self.instantiateList(for: .favorites)
         favoritesViewController.title = L10n.Titles.favorites
         favoritesViewController.tabBarItem = UITabBarItem(title: L10n.Titles.favorites, image: #imageLiteral(resourceName: "favorite"), tag: 2)
         
-        let archiveViewController: BaseListViewController = self.instantiateList(for: .archive)
+        let archiveViewController: ItemsViewController = self.instantiateList(for: .archive)
         archiveViewController.title = L10n.Titles.archive
         archiveViewController.tabBarItem = UITabBarItem(title: L10n.Titles.archive, image: #imageLiteral(resourceName: "archive"), tag: 3)
         
         return [listViewController, favoritesViewController, archiveViewController]
     }
     
-    private func instantiateList(for type: TypeOfList) -> BaseListViewController {
-        return BaseListViewController(notifier: self.dataProvider.notifier(for: type), dataProvider: self.dataProvider, userManager: self.userManager, themeManager: self.themeManager, type: type)
+    private func instantiateList(for type: TypeOfList) -> ItemsViewController {
+        return ItemsViewController(notifier: self.dataProvider.notifier(for: type), dataProvider: self.dataProvider, userManager: self.userManager, themeManager: self.themeManager, type: type)
     }
     
-    func instantiateList(for tag: Tag) -> BaseListViewController {
-        return BaseListViewController(notifier: self.dataProvider.notifier(for: tag), dataProvider: self.dataProvider, userManager: self.userManager, themeManager: self.themeManager, type: .myList)
+    func instantiateList(for tag: Tag) -> TaggedItemsViewController {
+        return TaggedItemsViewController(notifier: self.dataProvider.notifier(for: tag), dataProvider: self.dataProvider, userManager: self.userManager, themeManager: self.themeManager, tag: tag)
     }
     
     func instantiateSettings() -> SettingsViewController {
